requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-firebase==1.2
requests==2.10.0
simplejson==3.8.2
Werkzeug==0.11.10
